# docs/openapi/image-server.yaml
openapi: 3.0.3
info:
  title: Image Server API
  description: |
    Image upload and transformation service. Originals are stored behind a
    pluggable storage backend (local filesystem first); images are served back
    with on-the-fly transforms via query params, and derivatives are cached.

    This spec is the source of truth for the API. Handler interfaces are
    generated from it with oapi-codegen (make openapi-gen), and every
    request/response is validated against it in the pre-push API spec gate.
  version: 0.1.0
servers:
  - url: http://localhost:8080

security:
  - ApiKeyAuth: []

paths:
  /v1/images:
    post:
      operationId: uploadImage
      summary: Upload an image (multipart)
      description: |
        Uploads an image file. The real type is detected from magic bytes;
        JPEG, PNG, GIF, WebP, TIFF, and the HEIF family (HEIC, HEIF, AVIF, as
        produced by iPhones and modern Android cameras) are accepted.
        Byte-identical uploads are deduplicated by sha256 and return the
        existing record.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Image stored (or existing record returned on dedup)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /v1/images/from-url:
    post:
      operationId: uploadImageFromURL
      summary: Upload an image by URL
      description: |
        The server fetches the URL and stores the image. Only public http(s)
        URLs are accepted; requests resolving to private, loopback,
        link-local, or cloud-metadata addresses are rejected, including via
        redirects (SSRF guard).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  example: https://example.com/photo.jpg
      responses:
        "201":
          description: Image fetched and stored (or existing record on dedup)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /v1/images/{id}:
    get:
      operationId: getImage
      summary: Get an image, original or transformed
      description: |
        Without query params, returns the original. With params, returns a
        transformed derivative (generated once, then cached). Responses
        include Cache-Control headers.
      parameters:
        - $ref: "#/components/parameters/ImageID"
        - name: w
          in: query
          description: Target width in pixels
          schema:
            type: integer
            minimum: 1
        - name: h
          in: query
          description: Target height in pixels
          schema:
            type: integer
            minimum: 1
        - name: fmt
          in: query
          description: |
            Output format. The server only encodes to jpeg, png, or webp.
            Sources in a non-web format (HEIC, HEIF, AVIF, TIFF) are served
            back unchanged when no transform is requested, but transforming
            one requires an explicit fmt (a request without it returns 400).
          schema:
            type: string
            enum: [jpeg, png, webp]
        - name: q
          in: query
          description: Output quality for lossy formats
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: fit
          in: query
          description: "cover: crop to fill w x h. contain: fit within w x h."
          schema:
            type: string
            enum: [cover, contain]
        - name: strip
          in: query
          description: |
            Strip metadata (EXIF, including GPS, plus XMP/IPTC and comments)
            from the served image. The stored original is never modified.
            When combined with a resize or format change the strip is applied
            during that re-encode; on its own it is a lossless, byte-preserving
            removal for JPEG and PNG. Lossless strip is unsupported for other
            formats on its own and returns 415 (add fmt to strip via re-encode).
          schema:
            type: boolean
      responses:
        "200":
          description: Image bytes
          headers:
            Cache-Control:
              schema:
                type: string
              description: Caching policy for the (possibly derived) image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
            image/tiff:
              schema:
                type: string
                format: binary
            image/heic:
              schema:
                type: string
                format: binary
            image/heif:
              schema:
                type: string
                format: binary
            image/avif:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      operationId: deleteImage
      summary: Delete an image
      description: Deletes the original, cached derivatives, and metadata.
      parameters:
        - $ref: "#/components/parameters/ImageID"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /v1/images/{id}/meta:
    get:
      operationId: getImageMeta
      summary: Get image metadata
      parameters:
        - $ref: "#/components/parameters/ImageID"
      responses:
        "200":
          description: Image metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /healthz:
    get:
      operationId: healthz
      summary: Health check
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  parameters:
    ImageID:
      name: id
      in: path
      required: true
      description: Image ID
      schema:
        type: string
        format: uuid

  schemas:
    Image:
      type: object
      required:
        - id
        - original_filename
        - content_hash
        - mime_type
        - width
        - height
        - size_bytes
        - created_at
      properties:
        id:
          type: string
          format: uuid
        original_filename:
          type: string
        content_hash:
          type: string
          description: sha256 of the original bytes (hex)
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        mime_type:
          type: string
          example: image/jpeg
        width:
          type: integer
        height:
          type: integer
        size_bytes:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string

  responses:
    BadRequest:
      description: Invalid request (bad params, invalid image, blocked URL)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid X-API-Key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Image not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PayloadTooLarge:
      description: Upload exceeds the configured size limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: Not a supported image type (detected by magic bytes)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded for this API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"